shared/backend/requirements.txt
# ----------------------------
# ✅ Core Framework
# ----------------------------
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ----------------------------
# ✅ Database & ORM
# ----------------------------
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0

# ----------------------------
# ✅ Auth & Security
# ----------------------------
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pycryptodome==3.20.0

# ----------------------------
# ✅ Data & Scheduling
# ----------------------------
pandas==2.2.3
apscheduler==3.10.4

# ----------------------------
# ✅ Web & Utils
# ----------------------------
requests==2.32.3
python-dotenv==1.0.1
email-validator==2.2.0
pydantic==2.9.2
pydantic-settings==2.5.2

# ----------------------------
# ✅ Optional (if you have file upload routes)
# ----------------------------
python-multipart==0.0.9

# ----------------------------
# ✅ Production server
# ----------------------------

gunicorn==23.0.0